LayoutTests/fast/webgpu/nocrash/fuzz-147280474-expected.txt

CONSOLE MESSAGE: promises created
CONSOLE MESSAGE: the end
CONSOLE MESSAGE: fuzz-147280474.html
CONSOLE MESSAGE: Pass
